Update Main.c
Added alarm interrupt from the RTC chip.  On the interrupt a falling edge occurs on RA3 and the time is read whenever the alarm is set.  Also fixed the time to display 24 hour time format correctly.  This is the last version I will probably work on.  However, I plan on adding a schematic and PCB at some point.